Modificacion declaracion de función

+++ include/matematica.c
@@ -1,5 +1,5 @@
 #include <matematica.h>
 #include <stdlib.h>
 //manda a llamar a la función
-float operacion (char opcion, int num1, int num2);
+float operacion (char opcion, float num1, float num2);
 float conversion (char vector[],int tamaño);
